fastapi-db-tests-factoryboy/requirements.txt
fastapi==0.66.0
sqlalchemy==1.4.20
uvicorn[standard]==0.14.0
psycopg2-binary==2.9.1
black==21.6b0
requests==2.25.1
factory-boy==3.2.0
pytest-factoryboy==2.1.0
